Backport munit ports and fix recyclable assertion
